test(deepseek-v41): consolidate unit tests by component
